Move demangle impl to cpp
